Crate pixie_shared

Re-exports

pub use bijection::Bijection;

pub use config::*;

Modules

bijection

config

Structs

Address

Chunk

Describes one chunk from a disk.

ChunkStat

HintPacket

Image

An image is given by the list of chunks of the disk, the index of the boot entry that boots it, and the contents of that boot entry.

ImageStat

Station

Enums

Action

TcpRequest

UdpRequest

Constants

ACTION_PORT

BODY_LEN

CHUNK_SIZE

HEADER_LEN

PACKET_LEN

Functions

to_hex

Type Aliases

ChunkHash

The hash of a chunk of a disk.

Ip

Offset

The offset of the chunk of a disk.