Crate pixie_shared

Re-exports

pub use bijection::Bijection;

pub use config::*;

Modules

bijection

chunk_codec

config

util

Structs

Chunk

Describes one chunk from a disk.

ChunkStats

HintPacket

Image

An image is given by the list of chunks of the disk, the index of the boot entry that boots it, and the contents of that boot entry.

ImagesStats

RegistrationInfo

Enums

Action

StatusUpdate

TcpRequest

A request for the tcp server.

UdpRequest

A request for the udp server.

Constants

ACTION_PORT

CHUNKS_PORT

udp port for chunk broadcasting.

HINT_PORT

udp port for registration hint broadcasting.

MAX_CHUNK_SIZE

Maximum size in bytes for a chunk.

PING_PORT

udp port for pings from workstations.

UDP_BODY_LEN

The maximum number of bytes in a udp packet with mtu 1500

Type Aliases

ChunkHash

The hash of a chunk of a disk.

ChunksStats

Offset

The offset of the chunk of a disk.